mem0_practice/requirements.txt
openai
mem0ai
pymilvus
milvus-lite
